src/test/java/homework3/task3/UserRepositoryTest.java
package homework3.task3;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class UserRepositoryTest {

    private UserRepository userRepository;

    @BeforeEach
    void setUp() {
        userRepository = new UserRepository();
    }

    @AfterEach
    void tearDown() {
        userRepository = null;
    }

    @Test
    @DisplayName("Проверка метода добавления пользователя в общий список")
    void testAddUser() {
        User adminUser = new User("admin", "pass321", true);
        User regularUser = new User("user1", "qwerty", false);
        userRepository.addUser(adminUser);
        assertTrue(userRepository.findByName("admin"));

        userRepository.addUser(regularUser);
        assertTrue(userRepository.findByName("user1"));
    }

    @Test
    @DisplayName("Проверка метода поиска пользователя по логину в общем списке")
    void testFindByName() {
        User user = new User("testUser", "password", false);
        userRepository.addUser(user);

        assertTrue(userRepository.findByName("testUser"));
        assertFalse(userRepository.findByName("unanimous"));
    }

    @Test
    @DisplayName("Проверка метода разлогирования всех пользователей кроме администратора")
    void testLogOutAllExceptAdmins() {
        User adminUser = new User("admin", "pass321", true);
        User regularUser = new User("user1", "qwerty", false);
        userRepository.addUser(adminUser);
        userRepository.addUser(regularUser);

        userRepository.logOutAllExceptAdmins();

        assertTrue(adminUser.isAuthenticate);
        assertFalse(regularUser.isAuthenticate);
    }

    @Test
    @DisplayName("Проверка метода разлогирования всех пользователей")
    void testLogOutAll() {
        User adminUser = new User("admin", "pass321", true);
        User regularUser = new User("user1", "qwerty", false);
        userRepository.addUser(adminUser);
        userRepository.addUser(regularUser);

        userRepository.logOutAll();

        assertFalse(adminUser.isAuthenticate);
        assertFalse(regularUser.isAuthenticate);
    }


}